+++ lib/eleven_rb.rb
@@ -108,6 +108,7 @@ require_relative 'eleven_rb/resources/models'
 require_relative 'eleven_rb/resources/user'
 require_relative 'eleven_rb/resources/sound_effects'
 require_relative 'eleven_rb/resources/music'
+require_relative 'eleven_rb/resources/speech_to_speech'
 
 # High-level components
 require_relative 'eleven_rb/voice_slot_manager'